chore: rename index.tsx to ScopedVariables.tsx
