+++ src/components/scopedVariables/ScopedVariables.tsx
